requirements.txt

streamlit
pandas
python-dotenv
google-generativeai>=0.5.0